java: added/updated javadoc
The code contained some non javadoc conforming comments which were refactored
to valid javadoc. The added javadoc is derived from background knowledge
about the library and perceptions made from a code review. Because of the
absence of a specification the javadoc is based on assumptions how rxtx might
be intended to work. It can not replace these missing contracts.