impl-serial: hide driver class from API users
The driver implementation must not be available to API users
because they must not rely on implementation details. The
export was a huge mistake all the time. We might break some
implementations with this change! However I think it is
important for future rxtx versions. In the meantime
backports should mark the class as deprecated.
The driver is marked as final for performance reasons.