impl-serial: hide Zystem from API users
The Zystem class must not be available to API users
because they must not rely on implementation details. The
export was a huge mistake all the time. We might break some
implementations with this change! However I think it is
important for future rxtx versions. In the meantime
backports should mark the class as deprecated.