api: made driver event classes final
The API user must not be able to abuse event objects for own
purpose by subclassing, because when we add methods to these
classes they could colide with methods in the subclass.
This commit might break user code! Backports should only add
a note to the javadoc and not add the final modifier.