api: hide port event creation
The *PortEvent classes are intended to be created by drivers only.
However due to a design failure the constructors were expored in
the past. This is solved by this patch.

Now, the *PortEvent class' constructors are hidden to package
visibility. They are therefor no more available to the API user
which is good.

Because the drivers which are required to use these constructors
are also in a different package, they can't create PortEvents
directly. For this purpose the new class EventFactory was created.

EventFactory is a singleton factory class which creates
PortEvents on request. Although this class is public the API user
can not obtain instances of it by accident.

All drivers obtain the singleton instance of EventFactory via
their DriverContext instance. An instance of the new
DriverContext class is passed to every registered driver
(registration is currently hard coded) via the initialize()
method.

This patch contains factory methods for serial and parallel
port events. Other events are not exported because it is very
likely that this functionality will be removed from rxtx in
this review branch.