implemented driver discovery via ServiceLoader
Until now the instantiation of (the only available)
driver was hardcoded. Now the available CommDrivers
are discovered by the Java SE6 ServiceLoader mechanism.
This enables developers to plugin their own driver
implementations.

To control the discovery process the new class
DriverManager was introduced.

This patch changes functional backward compatibility
since the user is forced to call
DriverManager.loadDrivers() before any other interaction
with the API. This behaviour was introduced because
the old way to load the drivers from the static init
section of CommPortIdentifier, turned out to be
very test unfriendly. Additionally the time when
the drivers are initialized could not be controlled,
which results in bad responsiveness while the first
contact with the API.

This patch includes a test case for the new functionality.