implemented JAR aware library loader
Until now, loading native JNI libraries was done by calling
System.loadLibrary() directly in the static initialisation
section of CommPortIdentifier. System.loadLibrary() expected an
appropriate library to be on the classpath (or library path).
This approach has two problems:

1) With the introduction of 64bit systems it was no longer
ensured that the correct library for the system was loaded.
This is because System.loadLibrary() does not care about the
architecture of the available libraries at all.
It distinguishes between them using some naming conventions
for different platforms only (e.g. libfoo.so vs. foo.dll).
This is not sufficient to distinguish between e.g. a
linux x86_64 and x86 library, because they have the same
file name. This led to some hacks experimenting with name
suffixes, weaken the code quality.

2) Because users were forced to set up the library path, the
usage of the API was more difficult then necessary.

To overcome these limitations a LibraryLoader was created, which
can load JNI libraries from a JAR file and other sources. The
loader distinguishes between architecture and platform using
well defined postfixed library names. The postfixed name of the
library correlates with the artifact directory.

e.g. librxtxSerial-linux-x86.so is called POSTFIXED whereas
librxtxSerial.so is called NON-POSTFIXED

There are three basic use cases:

I) Out of the Box (load from resource)
The user obtains the pre-packed default JAR distribution of RXTX
which contains the default drivers and their respective JNI
libraries for all supported (or chosen) platforms.

Whenever the LibraryLoader can find the JNI libraries for the
current platform inside of the JAR (or registered driver JARs
when using additional drivers) these will be extracted to a
temporary directory and being loaded.

II) Redistribution (load from classpath)
Because the extraction from JAR at every program startup, might
be performance critical, a developer of a program depending on
RXTX might decide to setup the libraries once (e.g. via some
installer). In this case the libraries must not be packed to the
JARs.

When the LibraryLoader detects the absence of a library in the
JAR, it falls back to load it from the classpath (or library
path).

III) OS level installation (load from system library path)
When RXTX (and RXTX driver implementations) are distributed via
a system package manager (e.g. apt-get on debian based platforms)
the LibraryLoader does not need to select the correct binary.
This is because the installation package will set up only the
correct binary matching the platform. This is assured by the
system package manager itself. In such a case the long postfix
as used with I and II might be redundant and confusing. Therefor
it can be ommitted by the packager.

When the LibraryLoader can not find a postfixed version of a
library in the corresponding JAR, nor on the classpath, it
will look for a non-postfixed version on the system library
path. An application as described in II should never expose
non-postfixed libraries in its classpath. This could lead to
an incompatible JNI contract.