cleanup logging and porting to java logger
A LOT of debug System.out.println() statements cluttered up the
code and made it difficult to read.

Abusing the default output with trace messages is concidered to
be bad style. The important error and trace messages were therefor
ported to use the java built in logging facility. Less important
messages were removed.

In addition the logger approach possipbly eases bug reporting,
because tracing is possible without recompiling the library.