removed parallel port support
The default serial and parallel port drivers of RXTX were both
implemented in the RXTXCommDriver class. This is against the
intention of the SPI, which suggests separate driver classes.

Because the current development focuses on the serial port driver
as a reference implementation of the SPI, the parallel port
support is completely removed.

The parallel port might be reimplemented as a separate driver
in the future.