separated driver implementation from API
Until now rxtx was making no difference between API and
implementation. The whole java part was a single package.
This is an optimal incubator for spaghetti code and prevents
information hiding. Therefor the implementation of the
serial (and parallel) ports were moved to a new java
package gnu.io.impl.serial which is located in the new
artifact rxtx-impl-serial.

The separation makes clear that the
rxtx serial driver is only one possible implementation of
the API. This was forgotten in the past and the code
of serial, parallel and other drivers got cluttered.
Seperating the implementation assures, that the API is
not depending on the implementation.

This commit is still suboptimal because the serial and
parallel port driver implementations depend on eachother
and need to reside in the same package. In the future
the parallel port driver should be moved to another
package/artifact.