we're forced to make driver impl public again
In the review/visibility branch there was an attempt to hide
all implementation details from the API user. It turned out
this can not be easily archived in the case of the
implementations of CommDriver. To make the loosely coupled
API/implementation via reflection work, the implementation
class must be public and provide a public constructor. This
is also true for the planed ServiceLoader approach. Although
there are ways to modify the visiblity of classes using
reflection, this might be prohibited by the SecurityManager.
So we choose the way to make the implementation public.

To prevent the user from depending on the implementation details
by accident, we put on a huge DON'T DO THAT sign via the
javadoc deprecated tag.