libLoader: using unique TempFile numbering
The LibraryLoader extracts library binaries to the default
temporary location. These files are marked with the
File.deleteOnExit() method and are deleted when the JVM
is shutting down. To prevent problems when more than
one JVM instance is running, this fix uses the
File.createTempFile() method to enable unique file naming
for the extracted binaries.