add release build mechanism
The top level POM decides based on the property
${make.release} which build type to do.

1) development build
If maven is invoked by
$> mvn [-Pwith-some-extra-platform,...] install
a complete build cycle runs and bundles a rxtx for the given
platform(s).

2) release preparation
If maven is invoked by
$> mvn -Dmake.release=prepare install
only the platform independent parts of rxtx are build and
installed to the local maven repository.

3) release bundle
If maven is invoked by
$> mvn -Dmake.release=bundle install
a rxtx release bundle is created from the artifacts in
the local maven repository.

To create a release:
* run step 2) on machine A
* build all the native binaries by running "mvn install" in
their subdirectories. Some binaries can only compile on
certain operating systems (e.g. apple OSX), so you might
have to run these builds on different machines B, C.
* deploy all binaries to machine A's maven repository
* run step 3) on machine A