rxtxSerial: rename artifact directories
Due to the library naming conventions on *nix platforms,
some artifact names have a "lib" prefix, some do not. This is
okay, but their subdirectories have the prefix as well. This is
annoying for a matrix project setup in jenkins, because the
subprojects' POMs can't be located with a pattern.

This changeset renames the subdirectories, but keeps the "lib"
prefix in the artifact names.