change creation of JNI headers from deprecated javah to "javac -h"
Since we cannot use the convenient javah facility of native-maven-plugin,
the JNI headers are no longer generated by each rxtxSerial-* port artifact,
but as a .zip byproduct of the rxtxSerial-java artifact. This is achieved
by adding the additional "-h" flag during compilation of the java sources.
Since the native-maven-plugin does no longer provide the include paths to
the JNI headers, they are now added manually.