stop loading of rxtx property files from java.ext.dirs locations
The extension mechanism was removed from Java per JEP 220 targeting
Java 9. This resulted in a null java.ext.dirs property leading to
NullPointerException on such modern JVMs. Since the majority of users are
believed to never have used the extension mechanism anyway, this feature
is removed completely. Property files are still loaded from Classpath.

This is related to the fix in the debian maintainers' repository:
https://salsa.debian.org/java-team/rxtx/commit/b8f23b66208a41aef299b0f825b26d11774e0b47