maven: remove automatic binary selection by os
Resolution of maven bug https://issues.apache.org/jira/browse/MNG-4565
changed the behavior of profile activation triggers from logic OR to logic AND
without providing fallback behavior.

The new behavior leads to problems in release build for profiles which shall be
automatically included in releases, but also automatically included in a
development build on some build hosts for convenience reasons. Such binaries
are no longer included in the release build, which is fatal.

So we drop some convenience of develoment build here to ensure the sanity of
our release build.