maven: use profiles only for binary platform selection
Previously, profiles were used to minimize the selected modules in the
release=prepare and release=bundle use cases. This is no longer feasible if
we want to use the activeByDefault profile to warn the user about a missing
binary platform selection.

This will result in slightly more work in the release=prepare use case,
because we now also build the rxtxSerial and rxtxSerial-test modules which are
not needed in this step. But I think usability is more important here. And in
the long run, it might be a good idea to build and execute the tests anyway
(as soon as we have some :) .